F# API: Series Aggregation API fulfilled

+++ Polars.CSharp/Expr/Aggregation.cs
@@ -213,7 +213,7 @@ public partial class Expr : IDisposable,IEquatable<Expr>
     /// <param name="ddof">“Delta Degrees of Freedom”: the divisor used in the calculation is N - ddof, where N represents the number of elements. 
     /// By default ddof is 1.</param>
     /// <returns>A series which length is 1</returns>
-    public Expr Std(int ddof = 1) => new(PolarsWrapper.Std(CloneHandle(), ddof));
+    public Expr Std(byte ddof = 1) => new(PolarsWrapper.Std(CloneHandle(), ddof));
 
     /// <summary>
     /// Get the variance.
@@ -221,7 +221,7 @@ public partial class Expr : IDisposable,IEquatable<Expr>
     /// <param name="ddof">“Delta Degrees of Freedom”: the divisor used in the calculation is N - ddof, where N represents the number of elements. 
     /// By default ddof is 1.</param>
     /// <returns>A series which length is 1</returns>
-    public Expr Var(int ddof = 1) => new(PolarsWrapper.Var(CloneHandle(), ddof));
+    public Expr Var(byte ddof = 1) => new(PolarsWrapper.Var(CloneHandle(), ddof));
 
     /// <summary>
     /// Get the median value.
